worked on PR issues
