+++ .gitignore
@@ -2,3 +2,4 @@
 temp.py
 __pycache__/
 *.pyc
+*.txt
